lbvs-lib/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lbvs-lib"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "João Pedro Cardoso Barbosa", email = "joao.c.barbosa@grad.ufsc.br" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pycryptodome==3.20.0"
]

[project.urls]
Documentation = "https://github.com/João Pedro Cardoso Barbosa/lbvs-lib#readme"
Issues = "https://github.com/João Pedro Cardoso Barbosa/lbvs-lib/issues"
Source = "https://github.com/João Pedro Cardoso Barbosa/lbvs-lib"

[tool.hatch.version]
path = "src/lbvs_lib/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/lbvs_lib tests}"

[tool.hatch.envs.test]
dependencies = [
  "pytest"
]

[tool.coverage.run]
source_pkgs = ["lbvs_lib", "tests"]
branch = true
parallel = true
omit = [
  "src/lbvs_lib/__about__.py",
]

[tool.coverage.paths]
lbvs_lib = ["src/lbvs_lib", "*/lbvs-lib/src/lbvs_lib"]
tests = ["tests", "*/lbvs-lib/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

#[tool.hatch.build]
